Update main.py
I've added the option to choose a number of days from wich we can start.This option doesn't change the usual way of using the script. But if we want to start synchronizing from 30 days from now, we just have to add 30 at the end of the line. This is useful if we want to manually sync a date far away with wasting time.